Se modifica tipo de dato para el UDI

+++ Jfk.Tc.Decn.Domain/Entities/Usuario.cs
@@ -13,11 +13,11 @@ namespace Jfk.Tc.Decn.Domain.Entities
         public string NumeroDocumento { get; set; }
         public decimal Ingresos { get; set; }
         public decimal GastosFamiliares { get; set; }
-        public int CuotasFinancieras { get;}
+        public int CuotasFinancieras { get; }
         public int Edad { get; set; }
         public int FactorDeRiesgo { get; set; }
         public decimal TasaInteres { get; set; }
-        public decimal ValorUDI { get; set; }
+        public int ValorUDI { get; set; }
         public string HabitoDePago { get; set; }
         public string EstadoCedula { get; set; }
         public string Alertas { get; set; }
